+++ Interface/AddOns/PanzaUI/core.lua
@@ -77,6 +77,20 @@ local function configFrames()
 end
 
 --------------------------------------------------------------------------------
+-- CONFIGURE PERSONAL RESOURCE DISPLAY
+--------------------------------------------------------------------------------
+
+local function configPRD()
+    -- Change the status bar texture once during initialization.
+    PersonalResourceDisplayFrame.PowerBar:SetStatusBarTexture("Interface\\AddOns\\SharedMedia\\statusbar\\Wglass")
+
+    PersonalResourceDisplayFrame.PowerBar:HookScript("OnUpdate", function(self)
+
+    end)
+
+end
+
+--------------------------------------------------------------------------------
 -- CONFIGURE TOOLTIPS
 --------------------------------------------------------------------------------
 
@@ -198,6 +212,7 @@ end
 local function InitializeAddon()
 
     configFrames()
+    configPRD()
     configTooltips()
     setupCVars()